fix: revamp move operations to respect current filter
This fixes a bug where (say) a node was moved as child-of-previous ('>')
but the previous node in the model was hidden under the current filter
(e.g., completed), and then the node disappeared.

A less severe instance of this is when multiple move-up/down ops were
required with no visible effect b/c there were hidden nodes in between.

Now we have split up move operations: First find the "anchor" EID that
the user refers to, determine the target relative to that, and move
there.